[QOLDEV-863] ignore malformed Solr index archives

+++ recipes/solr-deploy.rb
@@ -272,7 +272,7 @@ bash "Copy latest index from EFS" do
         rsync -a --delete #{efs_data_dir}/ #{real_data_dir}/
         CORE_DATA="#{real_data_dir}/data/#{core_name}/data"
         LATEST_INDEX=`ls -dtr $CORE_DATA/snapshot.* |tail -1`
-        if (echo "$LATEST_INDEX" |grep "[.]tgz$" >/dev/null 2>&1); then
+        if (echo "$LATEST_INDEX" |grep "[.]tgz$" >/dev/null 2>&1) && (tar tzf "$LATEST_INDEX" >/dev/null); then
             mkdir -p "$CORE_DATA/index"
             rm -f $CORE_DATA/index/*; tar -xzf "$LATEST_INDEX" -C $CORE_DATA/index
         fi
